feat: Created Markdown Description component that can be utlized multiple places
