+++ package.json
@@ -87,7 +87,8 @@
     "vite-plugin-linter": "^2.0.2",
     "vite-plugin-static-copy": "^0.13.0",
     "vite-plugin-svgr": "^2.4.0",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "react-markdown": "^9.0.1"
   },
   "peerDependencies": {
     "@emotion/react": "^11.11.0",
@@ -138,7 +139,6 @@
     "@hyperplay/utils": "^0.1.0",
     "@tabler/icons-react": "^2.46.0",
     "@valist/sdk": "^2.9.1",
-    "ethereum-blockies-base64": "^1.0.2",
-    "react-markdown": "^9.0.1"
+    "ethereum-blockies-base64": "^1.0.2"
   }
 }
