+++ src/components/MarkdownDescription/MarkdownDescription.module.scss
@@ -1,0 +1,10 @@
+.description {
+  white-space: pre-wrap;
+  text-wrap: wrap;
+  line-height: var(--space-xl);
+  color: var(--color-neutral-400);
+}
+
+.linkBtn {
+  padding: 0 !important;
+}
