+++ src/components/QuestDetails/index.module.scss
@@ -63,7 +63,3 @@
   flex-direction: column;
   height: 100%;
 }
-
-.linkBtn {
-  padding: 0 !important;
-}
